Core fundamentals working.

+++ core/src/division/round.rs
@@ -1,2 +1,2 @@
 pub(crate) type RoundName = String;
-pub(crate) type RoundIndex = u64;
+pub(crate) type RoundIndex = usize;
